move connectors::Feature to context::Feature
